### 2022-06-22 #4497 @editorialbot
[PRE REVIEW]: The sspm R package: spatial surplus production models for the management of northern shrimp fisheries
**Submitting author:** @vlucet (<a href="http://orcid.org/0000-0003-0268-818X">Valentin Lucet</a>)
**Repository:** https://github.com/pedersen-fisheries-lab/sspm
**Branch with paper.md** (empty if default branch): main
**Version:** v0.9.1
