+++ DESCRIPTION
@@ -1,7 +1,7 @@
 Package: sspm
 Type: Package
 Title: Spatial Surplus Production Model Framework for Northern Shrimp Populations
-Version: 0.9.1.9000
+Version: 0.9.1.9001
 Authors@R: c(
     person("Valentin", "Lucet", email = "valentin.lucet@gmail.com", role = c("aut", "cre", "cph")),
     person("Eric", "Pedersen", email = "eric.pedersen@concordia.ca", role = c("aut")))
