### 2022-06-27 72e972b
attempt to fix fonya's plotting
