+++ pkgdown.yml
@@ -1,8 +1,8 @@
 pandoc: 2.7.3
-pkgdown: 2.0.4
+pkgdown: 2.0.5
 pkgdown_sha: ~
 articles:
   An_example_with_simulated_data: An_example_with_simulated_data.html
   Package_and_workflow_design: Package_and_workflow_design.html
-last_built: 2022-06-21T14:27Z
+last_built: 2022-06-27T15:05Z
 
