+++ vignettes/An_example_with_simulated_data.Rmd
@@ -202,7 +202,7 @@ sspm_model
 sspm_model_fit <- sspm_model %>% 
   spm(log_productivity ~ sfa +
         weight_per_km2_all_predators_lag_1 +
-        smooth_space(by = weight_per_km2_borealis_with_catch) +
+        smooth_space(by = weight_per_km2_borealis_with_catch_lag_1) +
         smooth_space(), 
       family = mgcv::scat)
 
