+++ pkgdown.yml
@@ -4,5 +4,5 @@ pkgdown_sha: ~
 articles:
   An_example_with_simulated_data: An_example_with_simulated_data.html
   Package_and_workflow_design: Package_and_workflow_design.html
-last_built: 2022-06-27T15:05Z
+last_built: 2022-06-27T21:08Z
 
