### 2022-07-04 85384e2
closes #42 plotting with only one boundary now works
