### 2022-07-04 232bb48
plotting only preds is now easier
