+++ vignettes/An_example_with_simulated_data.Rmd
@@ -248,5 +248,6 @@ plot(sspm_model_fit, biomass = "weight_per_km2_borealis", use_sf = TRUE)
 
 ```{r}
 plot(sspm_model_fit, biomass = "weight_per_km2_borealis",
-     next_ts = TRUE, aggregate = TRUE, scales = "free", interval = T)
+     next_ts = TRUE, aggregate = TRUE, scales = "free", 
+     smoothed_biomass = TRUE, interval = T)
 ```
