### 2022-07-06 #4497 @Bisaloo
Hi @hvoltbb wave, would you be interested in reviewing this [R package](https://github.com/pedersen-fisheries-lab/sspm) and [the associated paper](https://raw.githubusercontent.com/openjournals/joss-papers/joss.04497/joss.04497/10.21105.joss.04497.pdf) for JOSS? The first author of the submission recommended you as a reviewer.

> The sspm R package: spatial surplus production models for the management of northern shrimp fisheries

You can have a look at our [review criteria](https://joss.readthedocs.io/en/latest/review_criteria.html) and [review checklist](https://joss.readthedocs.io/en/latest/review_checklist.html) to get an idea of what is expected relaxed.
