### 2022-07-06 #4497 @Bisaloo
@VLucet, while I'm still thinking about it, could you check and fix the third reference please? It seems off.

> DFO. (2019). An assessment of Northern Shrimp (Pandalus borealis) in Shrimp Fishing Areas 4–6 and of Striped Shrimp (Pandalus montagui) in shrimp fishing area 4 in 2018 (Can. {Sci}. {Advis}. {Sec}. {Sci}. {Advis}. {Rep}. No. 2019/027; p. 23).
