+++ paper/paper.bib
@@ -242,7 +242,8 @@
 }
 
 @techreport{dfo_assessment_2019,
-	type = {Can. {Sci}. {Advis}. {Sec}. {Sci}. {Advis}. {Rep}.},
+	type = {techreport},
+	institution: {Canadian Science Advisory Secretariat (CSAS)},
 	title = {An assessment of {Northern} {Shrimp} (\textit{{Pandalus} borealis}) in {Shrimp} {Fishing} {Areas} 4–6 and of {Striped} {Shrimp} (\textit{{Pandalus} montagui}) in shrimp fishing area 4 in 2018},
 	number = {2019/027},
 	author = {{DFO}},
