+++ .github/workflows/draft-pdf.yml
@@ -2,6 +2,8 @@ on:
   push:
     paths:
       - paper/paper.md
+      - paper/paper.bib
+      - .github/workflows/draft-pdf.yml
 
 jobs:
   paper:
