### 2022-07-06 #4497 @Bisaloo
> Hi Hugo, Yes, I can do that. When is the deadline?

There is no hard deadline. We usually aim at completing reviews in 3-4 weeks but it's flexible.

Please also note that we encourage a more interactive process than most journals. Ideally, the reviews should be a discussion with possible quite a lot of back and forth. This is useful to avoid getting stuck on a problem early on. Within this framework, you should feel free to have a short look whenever possible, post some comments, and then come back to it later.

Please let me know if that would work for you.
