### 2022-07-06 #4497 @Bisaloo
> I see. So I need to post my comments to this thread, instead of uploading
it to a traditional editorial management website, right?

Right. It's even better if you can open issues in the [project repository](https://github.com/pedersen-fisheries-lab/sspm) for each of your points. It's much easier to keep track progress that way.
