### 2022-07-11 #4497 @haddonm
Hi @Bisaloo,  thanks for thinking of me as this software and article look interesting. Unfortunately, despite being officially retired, I am already over-committed preparing to present some work at an up-coming project meeting and then trying to meet some deadlines of my own. However, in another couple of months I ought to become more available so reviewing activities if another fishery modelling related package turns up.  Regards Malcolm.
