### 2022-07-13 #4497 @Bisaloo
> Hi @Bisaloo, thanks for thinking of me as this software and article look interesting. Unfortunately, despite being officially retired, I am already over-committed preparing to present some work at an up-coming project meeting and then trying to meet some deadlines of my own. However, in another couple of months I ought to become more available so reviewing activities if another fishery modelling related package turns up. Regards Malcolm.

No problem, thanks for your quick answer and best of luck on your various projects!

I would recommend to fill out [this form](https://joss.theoj.org/reviewer-signup.html) to be added to our [list of potential reviewers](https://bit.ly/joss-reviewers) so other JOSS editors are also aware you are willing to review in the future.
