+++ R/plot-helpers.R
@@ -33,7 +33,7 @@ plot_train_test <- function(x, scales){
 # -------------------------------------------------------------------------
 
 plot_productivity <- function(x, aggregate, interval, use_sf, page, nrow, ncol,
-                              log, scales){
+                              log, scales, point_size, line_size){
 
   boundary_col <- spm_boundary(x)
   resp <- spm_response(spm_formulas(x))
@@ -60,13 +60,14 @@ plot_productivity <- function(x, aggregate, interval, use_sf, page, nrow, ncol,
                      ncol = ncol, time_col = time_col, log = log,
                      scales = scales, color_profile = color_profile,
                      aggregate = aggregate, interval = interval,
-                     boundary_col = boundary_col)
+                     boundary_col = boundary_col, point_size = point_size,
+                     line_size = line_size)
 
 }
 
 plot_biomass <- function(x, biomass, biomass_origin, aggregate, interval,
                          use_sf, page, nrow, ncol, log, scales,
-                         next_ts, smoothed_biomass){
+                         next_ts, smoothed_biomass, point_size, line_size){
 
   # Check that biomass is a character
   checkmate::assert_character(biomass)
@@ -123,7 +124,8 @@ plot_biomass <- function(x, biomass, biomass_origin, aggregate, interval,
                      ncol = ncol, time_col = time_col, log = log,
                      scales = scales, color_profile = color_profile,
                      aggregate = aggregate, interval = interval,
-                     boundary_col = boundary_col)
+                     boundary_col = boundary_col, point_size = point_size,
+                     line_size = line_size)
 
 }
 
@@ -183,9 +185,12 @@ process_actual_biomass <- function(x, biomass_origin, biomass, patch_area_col,
 spm_plot_routine <- function(smoothed_data, var, use_sf, page, nrow, ncol,
                              time_col, log, scales, color_profile,
                              aggregate = FALSE, interval =  FALSE,
-                             boundary_col = NULL) {
+                             boundary_col = NULL, line_size, point_size) {
 
-  smoothed_data <- units::drop_units(smoothed_data)
+  browser()
+
+  smoothed_data <- units::drop_units(smoothed_data) %>%
+    dplyr::mutate(linesize = line_size, pointsize = point_size)
 
   if (log) {
     smoothed_data[[var]] <- log(smoothed_data[[var]])
@@ -209,11 +214,12 @@ spm_plot_routine <- function(smoothed_data, var, use_sf, page, nrow, ncol,
     base_plot <- ggplot2::ggplot(data = smoothed_data) +
       ggplot2::geom_line(ggplot2::aes(x = .data[[time_col]],
                                       y = .data[[var]],
-                                      color = .data$color)) +
+                                      color = .data$color),
+                         size = smoothed_data$linesize) +
       ggplot2::geom_point(ggplot2::aes(x = .data[[time_col]],
                                        y = .data[[var]],
                                        color = .data$color),
-                          cex = 0.8) +
+                          cex = smoothed_data$pointsize) +
       ggplot2::labs(y = the_title) +
       ggplot2::theme_light() +
       ggplot2::scale_color_manual(values = color_profile) +
