### 2022-08-10 f7dbad3
add 2 arguments to control CI and PI visibility
