+++ R/plot-helpers.R
@@ -33,7 +33,8 @@ plot_train_test <- function(x, scales){
 # -------------------------------------------------------------------------
 
 plot_productivity <- function(x, aggregate, interval, use_sf, page, nrow, ncol,
-                              log, scales, point_size, line_size){
+                              log, scales, point_size, line_size,
+                              show_PI, show_CI){
 
   boundary_col <- spm_boundary(x)
   resp <- spm_response(spm_formulas(x))
@@ -61,13 +62,15 @@ plot_productivity <- function(x, aggregate, interval, use_sf, page, nrow, ncol,
                      scales = scales, color_profile = color_profile,
                      aggregate = aggregate, interval = interval,
                      boundary_col = boundary_col, point_size = point_size,
-                     line_size = line_size)
+                     line_size = line_size,
+                     show_PI = show_PI, show_CI = show_CI )
 
 }
 
 plot_biomass <- function(x, biomass, biomass_origin, aggregate, interval,
                          use_sf, page, nrow, ncol, log, scales,
-                         next_ts, smoothed_biomass, point_size, line_size){
+                         next_ts, smoothed_biomass, point_size, line_size,
+                         show_PI, show_CI){
 
   # Check that biomass is a character
   checkmate::assert_character(biomass)
@@ -125,7 +128,8 @@ plot_biomass <- function(x, biomass, biomass_origin, aggregate, interval,
                      scales = scales, color_profile = color_profile,
                      aggregate = aggregate, interval = interval,
                      boundary_col = boundary_col, point_size = point_size,
-                     line_size = line_size)
+                     line_size = line_size,
+                     show_PI = show_PI, show_CI = show_CI)
 
 }
 
@@ -186,7 +190,8 @@ process_actual_biomass <- function(x, biomass_origin, biomass, patch_area_col,
 spm_plot_routine <- function(smoothed_data, var, use_sf, page, nrow, ncol,
                              time_col, log, scales, color_profile,
                              aggregate = FALSE, interval =  FALSE,
-                             boundary_col = NULL, line_size, point_size) {
+                             boundary_col = NULL, line_size, point_size,
+                             show_PI, show_CI) {
 
   smoothed_data <- units::drop_units(smoothed_data) %>%
     dplyr::mutate(linesize = line_size, pointsize = point_size)
@@ -256,17 +261,22 @@ spm_plot_routine <- function(smoothed_data, var, use_sf, page, nrow, ncol,
 
       }
 
-      base_plot <- base_plot +
-        ggplot2::geom_ribbon(
-          ggplot2::aes(x = .data[[time_col]],
-                       ymin = .data[[CI_lower_name]],
-                       ymax = .data[[CI_upper_name]],
-                       fill = .data$color), alpha = 0.5)  +
-        ggplot2::scale_fill_manual(values = color_profile) +
-        ggplot2::labs(fill = "Type")
+      if(show_CI){
+
+        base_plot <- base_plot +
+          ggplot2::geom_ribbon(
+            ggplot2::aes(x = .data[[time_col]],
+                         ymin = .data[[CI_lower_name]],
+                         ymax = .data[[CI_upper_name]],
+                         fill = .data$color), alpha = 0.5)  +
+          ggplot2::scale_fill_manual(values = color_profile) +
+          ggplot2::labs(fill = "Type")
+
+      }
 
       if (PI_lower_name %in% names(smoothed_data) &
-          PI_upper_name %in% names(smoothed_data)){
+          PI_upper_name %in% names(smoothed_data) &
+          show_PI){
         base_plot <- base_plot +
           ggplot2::geom_ribbon(
             ggplot2::aes(x = .data[[time_col]],
[…]
