+++ man/plot.Rd
@@ -23,7 +23,9 @@
   nrow = 2,
   ncol = 2,
   log = FALSE,
-  scales = "fixed"
+  scales = "fixed",
+  show_PI = TRUE,
+  show_CI = TRUE
 )
 
 \S4method{plot}{sspm_fit,missing}(
@@ -44,7 +46,9 @@
   nrow = 2,
   ncol = 2,
   log = FALSE,
-  scales = "fixed"
+  scales = "fixed",
+  show_PI = TRUE,
+  show_CI = TRUE
 )
 }
 \arguments{
@@ -65,7 +69,7 @@ size.}
 \item{use_sf}{\strong{[logical]} Whether to produce a spatial plot.}
 
 \item{interval}{\strong{[logical]} (For sspm_fit & sspm_dataset) Whether to plot
-CI and Pi intervals.}
+CI and PI intervals.}
 
 \item{page}{The page to draw}
 
@@ -79,6 +83,10 @@ CI and Pi intervals.}
 \item{scales}{Are scales shared across all facets (the default,
 \code{"fixed"}), or do they vary across rows (\code{"free_x"}),
 columns (\code{"free_y"}), or both rows and columns (\code{"free"})?}
+
+\item{show_PI}{\strong{[character]} Whether to show the PIs.}
+
+\item{show_CI}{\strong{[character]} Whether to show the CIs.}
 
 \item{train_test}{\strong{[logical]} (For sspm_fit) Whether to plot a train/test
 pair plot.}
