### 2022-08-16 #4497 @Bisaloo
Hi @gmcdonald-sfg :wave:, I found your profile via your work on the [online book about fisheries analysis with R](https://sfg-ucsb.github.io/fishery-manageR/).

Would you be interested in reviewing this [R package](https://github.com/pedersen-fisheries-lab/sspm) and [the associated paper](https://raw.githubusercontent.com/openjournals/joss-papers/joss.04497/joss.04497/10.21105.joss.04497.pdf) for JOSS?

> The sspm R package: spatial surplus production models for the management of northern shrimp fisheries

You can have a look at our [review criteria](https://joss.readthedocs.io/en/latest/review_criteria.html) and [review checklist](https://joss.readthedocs.io/en/latest/review_checklist.html) to get an idea of what is expected :relaxed:.

Please feel free to suggest another member of your team if you're busy or you feel they might be a better fit.
