+++ README.Rmd
@@ -14,7 +14,7 @@ knitr::opts_chunk$set(
 ```
 
 # sspm
-<img src='man/figures/logo.png' align="right" height="150" width="130"/>
+<img src='man/figures/logo.png' style='text-align: right;' height="150" width="130"/>
 
 <!-- badges: start -->
 [![License: MIT](https://img.shields.io/badge/License-MIT-yellow.svg)](https://opensource.org/licenses/MIT/)
