+++ man/accessors-methods-sspm_boundary.Rd
@@ -85,10 +85,6 @@ spm_patches_area(object) <- value
 \item{sspm_object}{\strong{[sspm_boundary]} An object of class
 \link[=sspm_boundary-class]{sspm_boundary}.}
 
-\item{object}{
-    object from which to extract element(s) or in which to replace element(s).
-  }
-
 \item{value}{typically an array-like \R object of a similar class as
     \code{x}.}
 }
