+++ man/sspm-accessors-methods.Rd
@@ -86,10 +86,6 @@ spm_unique_ID(object) <- value
 \item{sspm_object}{\strong{[sspm OR adjacent]} An object of class
 \link[=sspm-class]{sspm} or others derivative classes.}
 
-\item{object}{
-    object from which to extract element(s) or in which to replace element(s).
-  }
-
 \item{value}{typically an array-like \R object of a similar class as
     \code{x}.}
 }
