+++ man/accessors-methods-discret_method.Rd
@@ -25,10 +25,6 @@ method_func(object) <- value
 \item{sspm_object}{\strong{[discretization_method]} An object of class
 \link[=discretization_method-class]{discretization_method}.}
 
-\item{object}{
-    object from which to extract element(s) or in which to replace element(s).
-  }
-
 \item{value}{typically an array-like \R object of a similar class as
     \code{x}.}
 }
