+++ man/accessors-methods-sspm_dataset.Rd
@@ -133,6 +133,9 @@ is_mapped(object) <- value
 \item{sspm_object}{\strong{[sspm_dataset]} An object of class
 \link[=sspm_dataset-class]{sspm_dataset}.}
 
+\item{object}{\strong{[sspm_dataset]} An object of class
+\link[=sspm_dataset-class]{sspm_dataset}.}
+
 \item{value}{typically an array-like \R object of a similar class as
     \code{x}.}
 }
