+++ pkgdown.yml
@@ -1,8 +1,8 @@
-pandoc: 2.7.3
-pkgdown: 2.0.5
+pandoc: 2.14.2
+pkgdown: 2.0.6
 pkgdown_sha: ~
 articles:
   An_example_with_simulated_data: An_example_with_simulated_data.html
   Package_and_workflow_design: Package_and_workflow_design.html
-last_built: 2022-07-06T14:52Z
+last_built: 2022-08-22T20:51Z
 
