+++ NEWS.md
@@ -1,3 +1,8 @@
+# sspm 0.9.2
+
+* Improved plotting, stability and user experience
+* Conformed to new HTML requirements as per required by CRAN
+
 # sspm 0.9.1.9000
 
 * Added a `NEWS.md` file to track changes to the package.
