### 2022-08-24 #4497 @eric-pedersen
Hi @Bisaloo,

I  would suggest James Thorson (user name James-Thorson), Sean Anderson (seananderson), Eric Ward (ericward-noaa), or Derek Ogle (droglenc), all of whom are experts in fisheries modelling and spatiotemporal models (as well as R code development).

As a note about any potential conflicts of interest: James Thorson and I are both members of the same working group on hierarchical modelling of ecological communities, but have not collaborated beyond that, and have only met indirectly via those working group meetings. Sean Anderson and I were both working at Fisheries and Oceans Canada at the same time and have corresponded before, but we have not collaborated on any projects or co-supervised any students together.
