### 2022-08-25 #4497 @Bisaloo
@eric-pedersen, amazing! Thanks for the suggestions and the details around possible conflict of interests :+1:
