### 2022-08-25 #4497 @Bisaloo
Hi @seananderson :wave:, would you be interested in reviewing this [R package](https://github.com/pedersen-fisheries-lab/sspm) and [the associated paper](https://raw.githubusercontent.com/openjournals/joss-papers/joss.04497/joss.04497/10.21105.joss.04497.pdf) for the Journal of Open Source Software (JOSS)?

> The sspm R package: spatial surplus production models for the management of northern shrimp fisheries

You were recommended as a reviewer by one of the authors and I agree you would be a great fit to review this submission.

You can have a look at our [review criteria](https://joss.readthedocs.io/en/latest/review_criteria.html) and [review checklist](https://joss.readthedocs.io/en/latest/review_checklist.html) to get an idea of what is expected :relaxed:.
