### 2022-08-29 #4497 @seananderson
Hi @Bisaloo, thanks for the invitation. I'd be keen to review this, but realistically, I'm off or travelling and busy starting Wednesday through most of September. The package looks interesting and it seems like a lot of work has gone into it. I do think I'd struggle to review the paper in its current form. There's a reproducible vignette, which is great, but I think I'd struggle to understand exactly what is happening and why (e.g., model equations, descriptions of what the various options do and why you might choose them). I think a little more hand holding of a new user could go a long way. :slightly_smiling_face:  I appreciate the equations may be in the in-press CSAS Research Document, which isn't publicly available yet to my knowledge.

The reviewer suggestions above look great. Some other ideas:
- [Brian Stock](https://brianstock.github.io/) (GitHub: brianstock)
- [John Best](https://www.posdefecology.net/) (GitHub: jkbest2)
- Authors of [this](https://academic.oup.com/icesjms/article/78/9/3265/6386085) paper on a spatial surplus production model. Raphael McDonald (GitHub RaphMcDo), [Dave Keith](https://scholar.google.com/citations?hl=en&user=rWQI9AYAAAAJ&view_op=list_works&sortby=pubdate) (herein I can't figure out or don't know people's GitHub usernames), ...
