### 2022-09-01 #4724 @editorialbot
[REVIEW]: The sspm R package: spatial surplus production models for the management of northern shrimp fisheries
**Submitting author:** @VLucet (<a href="http://orcid.org/0000-0003-0268-818X">Valentin Lucet</a>)
**Repository:** https://github.com/pedersen-fisheries-lab/sspm
**Branch with paper.md** (empty if default branch): main
**Version:** v1.0.0
