### 2022-09-01 #4724 @Bisaloo
:wave: :wave: :wave: @VLucet, @quang-huynh, @kellijohnson-NOAA this is the review thread for the paper. All of our communications will happen here from now on.

```@editorialbot generate my checklist```
