### 2022-09-23 #4724 @quang-huynh
I reviewed the software by installing the package on my local desktop and going through the example in the **An_example_with_simulated_data** vignette. I was able to replicate the steps in the vignette on my R console and replicate the figures.

**Documentation suggestions:**

I noticed that many figures in the vignette are too small, so the figure width needs to be wider in the Rmarkdown document.

I also see that the smoothened density estimates (step 8) are identical in all four panels. Is this correct? Do P1-4 represent SFA?

The package is well organized for those familiar with S4 class system. I would recommend reducing the number of methods for generics whenever possible but this is not a major concern. For example, `spm_as_dataset` has 4 defined methods but two should suffice (for `data.frame` and `sf`).

Are model diagnostic tools available to evaluate goodness of fit? I see that `sspm_model_fit@fit` is a bam object with various methods already defined in the `mgcv` and `stats` packages. The vignette can be updated to indicate where the model object is located and what diagnostics are available.

A summary method for the `sspm_fit` S4 class would be very useful for reporting estimates of biomass and productivity (both by time/area), and all other outputs that would be important to advise management.

***Functionality documentation*** The help documentation can be improved without much additional effort. For example, the help for `spm_as_boundary` says that the function creates a sspm_boundary object which is not helpful. The function divides polygons into smaller patches. An additional sentence to describe what these functions actually do should be added.

***Community guidelines*** A quick statement in the README.md to recommending users to report bugs in Github issues will meet this requirement.

**Paper suggestions:**

***State of the field*** There's not much reference to similar software. There are various software packages, e.g., spict, aspic, jabba, for surplus production models that estimate intrinsic rate of increase and carrying capacity. These model a unit stock without spatial considerations. However, sspm does not appear to estimate the inherent productivity of a population. Rather, sspm appears to use GAMs to estimate surplus production and productivity `(C_y + B_y)/C_y-1` over space and time, as well as with any covariates provided to the model (I am not currently able to fully review the methodology underlying the package as I don't have access to the Pedersen et al. Res Doc).

In a fisheries context, sspm reminds me more of the VAST and sdmTMB packages that implement spatiotemporal GLMMs to predict response variables across a spatial field over time. It may be possible to somewhat replicate sspm in sdmTMB, but estimation in sspm will be much, much faster (compared to the use of Gaussian Markov random fields in the other packages), and sspm conveniently addresses a specific use-case for fisheries assessment. I recommend reference and comparison of these packages with sspm.

***References*** The Pedersen et al and Prager reference needs to be edited to meet this requirement.

**Other paper suggestions:**

Line 13: Fisheries managers typically do not use analytical methods but receive advice from analysts. Perhaps "fisheries analysts" should be used instead of 'fisheries managers'?
Line 50: Suggest "Advice to managers advice should account for varying productivity"
Line 79: Should be "surplus production"
