### 2022-10-03 pedersen-fisheries-lab/sspm#53 @kellijohnson-NOAA
No link in "see for more details"
https://github.com/pedersen-fisheries-lab/sspm/blob/35232f970281ff137bb76835fb9f83149cc6391e/vignettes/Package_and_workflow_design.Rmd#L27

The previous code renders to
```
4. The first proper modelling step is to smooth the biomass and predictors data
by combining a sspm_dataset, and a sspm_discrete_boundary.
The user specifies a gam formula with custom smooth terms (see for more details).
The output is still a sspm_dataset object with a smoothed_data slot which contains
the smoothed predictions for all patches.
```
where there is no link in the 4th line of the text because `\autoref{tab:formula}` is not rendering properly.

This issue is part of a review for JOSS, see openjournals/joss-reviews#4724.
