### 2022-10-05 pedersen-fisheries-lab/sspm#55 @Bisaloo
Expand README
The README is likely to be the first point of contact between your package and potential users. It needs to have some examples of what can be done with your package.

You can get a more detailed recommendation for a README structure from [rOpenSci's devguide section on READMEs](https://devguide.ropensci.org/building.html#readme).

This is part of https://github.com/openjournals/joss-reviews/issues/4724.
