### 2022-10-05 pedersen-fisheries-lab/sspm#57 @Bisaloo
Increase test coverage
The test coverage of this package is rather low. While high test coverage doesn't necessarily guarantee that a package is well tested, low coverage does indicate that some sections of the code are tested at all.

[![Codecov test coverage](https://codecov.io/gh/pedersen-fisheries-lab/sspm/branch/main/graph/badge.svg)](https://app.codecov.io/gh/pedersen-fisheries-lab/sspm)

There is no strict threshold above which coverage is usually considered sufficient but I think you should aim for at least for 80%. It happens that some parts of the code can be more difficult to test. Feel free to let me know if that's the case and I can assist you.

This is part of https://github.com/openjournals/joss-reviews/issues/4724.
