### 2022-10-11 #4497 @Bisaloo
Hi @haddonm, sorry about the spam. It's not really a mistake but an unfortunate GitHub feature. Since I tagged you earlier in this thread, you have been automatically subscribed to this issue and now receive a notification each time a new comment is added to this thread (see the yellow rectangle in the screenshot attached).
You can unsubscribe but the good news is that all discussion has now moved to #4724 so even if you don't manage to unsubscribe, you should not be bothered again.

![image](https://user-images.githubusercontent.com/10783929/195019851-cee48e71-ec60-44e6-ac8b-92a42630678a.png)
