### 2022-11-20 a0bd117
add commented structure to get started
