+++ _pkgdown.yml
@@ -1,3 +1,13 @@
 template:
   params:
     bootswatch: flatly
+# reference:
+# - title: Layers
+# - subtitle: Geoms
+#   desc: Geom is short for geometric element
+# - contents:
+#   - starts_with("geom")
+# - subtitle: Stats
+#   desc: Statistical transformations transform data before display.
+#   contents:
+#   - starts_with("stat")
