+++ DESCRIPTION
@@ -16,7 +16,7 @@ BugReports: https://github.com/pedersen-fisheries-lab/sspm/issues
 URL: https://pedersen-fisheries-lab.github.io/sspm/
 Encoding: UTF-8
 Roxygen: list(markdown = TRUE)
-RoxygenNote: 7.2.1
+RoxygenNote: 7.2.2
 Depends: 
     sf,
     mgcv,
