### 2022-11-20 ea6792e
add internal keywords to some methods
