### 2022-11-20 7aefbd9
improves readme, closes #55
