### 2022-11-20 pedersen-fisheries-lab/sspm#54 @VLucet
- [ ] 8-14- The "Statement of need" is currently written as a list of bullet points but a paragraph that connects the dots for readers would be more useful. Or, at least better organization of the bullets is needed. For example 1 and 3 are connected but it is hard to understand their differences when listed as separate bullets. The mentioning of the example of Northern Shrimp as the second bullet distracts from the need for the package. I would place this last. The acronym SPM is not defined in this section nor is SSPM. Lastly, I will say that the need for population models is not stated. Why does a fisheries analyst or fisheries manager even need a population model in the first place?
- [ ] 18- remove "environmental"
- [ ] 19- inconsistent use of "such as", here it is in brackets and just two sentences later it is sectioned off using a comma. Either is fine but consistency is better. I would choose the latter.
- [ ] 21- remove comma
- [ ] 22-24- Remove the last sentence of this paragraph because it is essentially a repeat of the previous sentence.
- [ ] 27- comma after "parameters" if you choose to encapsulate "such as" this way, see note above.
- [ ] 30- add "are" before "still"
- [ ] 33- remove comma
- [ ] 35- remove "(i.e. surplus production)" because the previous line says nothing about surplus biomass or recruits
- [ ] 36- remove "and sex" because you can have a well-informed stock assessment model with unsexed age information
- [ ] 37- remove the comma and "age- or length-structure" should be "age or length structure"
- [ ] 38- text should be present tense, not past, and thus, "assumed" should be "assume". I am uncertain if this is the best location for the sentence that starts with "Classically," The sentence does not really pertain to the previous one about how SPMs have been used in data-poor situations. I think you could frame it as such though which would make it fit here a lot better. Otherwise, I think it needs to be the second sentence in this paragraph.
- [ ] 39- remove the comma before "and" and place a comma after "and" and after "such". Also, "a limited tool" is difficult to understand. I would change to something more like "having limited use".
- [ ] 41-44- These two sentences should in my opinion be grouped with the last sentence in the previous paragraph because they are talking about similar things. Furthermore, I think that the sentences would be more understandable if they were condensed into a single sentence, e.g., "SPMs also assume that populations are only affected by past abundance and fishing, which ignores changes that affect growth rates independently of fishing pressure, such as climate change."
- [ ] 44- Remove "One example is that of the"
- [ ] 47- Add ", is a prime example of a population thought to be affected by environmental conditions" after "(DFO, 2019)". Change "stocks" to "populations"
- [ ] 48- Remove the comma.
- [ ] 49- Remove "in the region".
- [ ] 49-52- I think that this sentence is a great candidate for an opening sentence of this paragraph rather than a closing sentence. Especially if you agree with my comment that 41-44 should be part of the previous paragraph.
- [ ] 53- Change "two" to "the two following". Add "models" after "statistical".
- [ ] 58- It is unclear what you mean by "This", please be more clear by removing "This" and providing a more in-depth description of what you are referring to.
- [ ] 59- remove the comma
- [ ] 53-67- It is unclear if you are using process-based or statistical models. After reading through a few times I now understand that you are using the latter but it should be made more clear such that readers can comprehend what you are using upon first read.
- [ ] 67- I am uncertain, but I think that it should "an R package".
- [ ] 69- "The" should be "the"
- [ ] 70- "simpler" is a relative term and I suggest being more explicit here.
- [ ] 72- You do not estimate the model, which is what is written.
- [ ] 73-74- I do not know what "manage biomass and harvest data" means.
- [ ] 74-75- Change the sentence starting with "Those" to "Those features are organized in the following stepwise workflow (Figure 1):
- [ ] 68-75- This paragraph makes it seem like the package creation was an after thought instead of the goal being to design an R package that can accommodate spatial data while fitting SSPMs. "and apply to any spatially structured stock" is grammatically incorrect. It seems like the sentence starting on line 62 through the end of that paragraph pertains more to this paragraph than where it is currently located. I would change where the paragraphs are separated and make sure that information is not redundant.
- [ ] 76-83- I recommend using this list as the larger outline for the list in the Package design section. It would be helpful for users to know what the functions listed in the Package design pertain to. That is, disperse the relevant topics that highlight R functions and output (1-8) under the headings 1-6 to create a nested list in the Package design section and no list in the Summary.
- [ ] 78- Place a comma before "or"
- [ ] 81- Is is really data or are they estimates?
- [ ] 83- I argue that you are not making forecasts but instead you are creating projections.
- [ ] 84- Change "suitable" to "suited"
- [ ] 87-88- The sentence starting with "The different ..." was not helpful to me and should be removed.
- [ ] 90-114- It is difficult to understand what is a function and what is a class because they are named similarly. Please be sure to use "()" after a function name or function call.
- [ ] 90-92- The first sentence does not seem like a complete sentence to me and I think that the two sentences should be combined, e.g., "Provided boundary data in the form of a shape file is converted into a sspm_boundary object using spm_as_boundary() to define the boundary/region of interest."
- [ ] 97-98- Change "(i.e." to "(i.e.,"
- [ ] 99- Add "()" after "spm_smooth" because it is a function name. "spatiotemporal" here is a bit misleading because you are just running discrete GAMS per group not linking the GAMs based on physical proximity like a full on spatial model would do, i.e., there are no Gaussian random fields that associate one point in space to the physical location of near and far points.
- [ ] 99- Change "GAM models" to "GAMs"
- [ ] 103- Add "()" after "spm_aggregate_catch" and remove comma
- [ ] 105- remove the comma
- [ ] 106- Add "()" after "sspm"
- [ ] 114- "and plots with the plot() method" is not a complete sentence and it should be because you are using ", and" which means that everything after "and" should be able to stand alone.
- [ ] Figure 1- The caption does not explain what the colors or shapes mean. I appreciate trying to make the infographic fit in the space available but the current design distracts from the comprehension of the figure because left to right or up and down do not mean anything. It would be nice if the left--hand side or the top represented all raw inputs and you could see the flow more easily on where the inputs go and exactly what outputs move on and what outputs need more post-processing before moving on.
- [ ] 116- Change "canadian Department of Fisheries of Oceans" to "Fisheries and Oceans Canada"
- [ ] 117- Change "fisheries" to "Fisheries", remove the comma before "and", and change "canadian" to "Canadian".
- [ ] 128-129- I am uncertain, but I think that white papers such as this should be treated as book titles where all significant words are capitalized.
Thanks so much for the review @kellijohnson-NOAA . My apologies for the long delay. I have just copied your comments and turned them into check boxes so I can do proper task management.

- [ ] 125-126- The title should be in lower case except "The".
- [ ] 129- Pandalus montagui should be in italics.
- [ ] 138- Please remove "fsz048" and add the relevant information regarding Volume 76, Issue 5, 09-10 2019, Pages 1305–1315 to the citation.
- [ ] 140-141- Please remove the curly braces. I am uncertain, but I think that white papers such as this should be treated as book titles where all significant words are capitalized. Please check if this item has a publication and page numbers.
- [ ] 147- This references is missing the Journal, page numbers, and a doi.
- [ ] 122-173- Please check the format of the references.
