+++ paper/paper.md
@@ -70,7 +70,7 @@ The key workflow steps are:
 
 1. Delineation of the boundary of the region of interest for the model. Boundary data is provided as a shapefile and converted into a `sspm_boundary` object with a call to `spm_as_boundary()`.
 2. The region within the boundary is discretized into patches with the `spm_discretize()` function, creating a `sspm_discrete_boundary` object.
-3. The `spm_as_dataset()` function turns user-provided data frames of raw observations into `sspm_dataset` objects that explicitly track locations, data types, and aggregation scales for each input. `sspm` recognizes three types of data: **trawl** (i.e. biomass estimates from scientific surveys), **predictors**, and **catch** (i.e. harvest). 
+3. The `spm_as_dataset()` function turns user-provided data frames of raw observations into `sspm_dataset` objects that explicitly track locations, data types, and aggregation scales for each input. `sspm` recognizes three types of data: **trawl** (i.e. biomass estimates from scientific surveys), **predictors**, and **catch** (i.e., harvest). 
 4. The `spm_smooth` function use spatiotemporal GAM models to smooth the biomass and predictor data, based on the spatial structure from `sspm_discrete_boundary`. The user specifies a GAM formula with custom smooth terms. The output is another `sspm_dataset` object with a `smoothed_data` slot which contains the smoothed predictions for all patches.
 5. The `spm_aggregate_catch` function aggregates catch into patches and years, and calculates patch-specific productivity for each year as the ratio of estimated biomass density plus catch from the next year, divided by estimated biomass density of the current year. The result is returned as a `sspm_dataset`.
 6. The `sspm` function combines productivity and predictor datasets into a single dataset. Additionally, the user may create lagged versions of predictors with `spm_lag()` and split data into testing and training sets for model validation with `spm_split()` at this stage.
