### 2022-11-26 7afa766
closes #92 + addresses a point brought up in the issue above
