+++ paper/paper.bib
@@ -108,7 +108,7 @@
 }
 
 @article{crowder_impacts_2008,
-	title = {The {Impacts} of {Fisheries} on {Marine} {Ecosystems} and the {Transition} to {Ecosystem}-{Based} {Management}},
+	title = {The {impacts} of {fisheries} on {marine} {ecosystems} and the {transition} to {ecosystem}-{based} {management}},
 	volume = {39},
 	url = {https://doi.org/10.1146/annurev.ecolsys.39.110707.173406},
 	doi = {10.1146/annurev.ecolsys.39.110707.173406},
