### 2022-11-26 e2f1f16
closes #97 closes #100
