+++ paper/paper.bib
@@ -244,7 +244,7 @@
 @techreport{dfo_assessment_2019,
 	type = {techreport},
 	institution = {Canadian Science Advisory Secretariat (CSAS)},
-	title = {An assessment of {Northern} {Shrimp} (\textit{{Pandalus} borealis}) in {Shrimp} {Fishing} {Areas} 4–6 and of {Striped} {Shrimp} (\textit{{Pandalus} montagui}) in shrimp fishing area 4 in 2018},
+	title = {An Assessment of {Northern} {Shrimp} (Pandalus borealis) in {Shrimp} {Fishing} {Areas} 4–6 and of {Striped} {Shrimp} (Pandalus montagui) in Shrimp Fishing Area 4 in 2018},
 	number = {2019/027},
 	author = {{DFO}},
 	year = {2019},
