+++ paper/paper.bib
@@ -29,7 +29,8 @@
 	editor = {Maravelias, Christos},
 	month = apr,
 	year = {2019},
-	pages = {fsz048},
+	pages = {1305--1315},
+	volume = {76},
 	file = {Karp et al. - 2019 - Accounting for shifting distributions and changing.pdf:/home/vlucet/Zotero/storage/JP6BDALA/Karp et al. - 2019 - Accounting for shifting distributions and changing.pdf:application/pdf},
 }
 
