+++ paper/paper.bib
@@ -253,9 +253,8 @@
 }
 
 @techreport{pedersenNewSpatialEcosystembased2021,
-	type = {Can. {Sci}. {Advis}. {Sec}. {Res}. {Doc}.},
-	title = {A new spatial ecosystem-based surplus production model for {SFA} 4-6 {Northern} {Shrimp}},
-	number = {2022/nn},
+	type = {Canadian Science Advisory Secretariat (CSAS) Research Document},
+	title = {A New Spatial Ecosystem-Based Surplus Production Model for {SFA} 4-6 {Northern} {Shrimp}},
 	author = {Pedersen, Eric J. and Skanes, Katherine and Le Corre, Nicolas and Koen-Alonso, Mariano and Baker, Krista},
 	year = {2022},
 	pages = {in press},
