+++ paper/paper.bib
@@ -147,10 +147,12 @@
 
 @article{prager_suite_1994,
 	title = {A suite of extensions to a nonequilibrium surplus-production},
+	journal = {Fishery Bulletin - National Oceanic and Atmospheric Administration},
 	language = {en},
 	author = {Prager, Michael H},
 	year = {1994},
-	pages = {17},
+	volume = {92},
+	pages = {374--389},
 	file = {Prager - A suite of extensions to a nonequilibrium surplus-.pdf:/home/vlucet/Zotero/storage/4HHXRZ4Y/Prager - A suite of extensions to a nonequilibrium surplus-.pdf:application/pdf},
 }
 
