### 2022-11-29 e8ff4b3
closes #107
