### 2022-11-29 pedersen-fisheries-lab/sspm#54 @VLucet
@eric-pedersen please check #113 and #115
