+++ paper/paper.bib
@@ -204,7 +204,6 @@
 	file = {Snapshot:/home/vlucet/Zotero/storage/L24AL4DI/a-primer-of-ecology-9780878933181.html:text/html},
 }
 
-
 @article{thorson_importance_2015,
 	title = {The importance of spatial models for estimating the strength of density dependence},
 	volume = {96},
@@ -219,7 +218,6 @@
 	year = {2015},
 	pages = {1202--1212},
 }
-
 
 @book{walters_fisheries_2004,
 	address = {Princeton, N.J., U.S.A.},
@@ -259,10 +257,8 @@
 	title = {A New Spatial Ecosystem-Based Surplus Production Model for {SFA} 4-6 {Northern} {Shrimp}},
 	author = {Pedersen, Eric J. and Skanes, Katherine and Le Corre, Nicolas and Koen-Alonso, Mariano and Baker, Krista},
 	year = {2022},
-	pages = {in press},
+	url = {https://www.dfo-mpo.gc.ca/csas-sccs/Publications/ResDocs-DocRech/2022/2022_062-eng.html}
 }
-
-
 
 @article{plaganyi2014Multispecies,
 	title = {Multispecies fisheries management and conservation: tactical applications using models of intermediate complexity},
@@ -279,5 +275,3 @@
 	year = {2014},
 	pages = {1--22},
 }
-
-
