### 2022-12-07 pedersen-fisheries-lab/sspm#54 @VLucet
All comments have been adressed, and subsequent issues will be closed when PR is closed
