### 2022-12-07 d1ddf25
fix plot sizes, closes #118
