+++ vignettes/An_example_with_simulated_data.Rmd
@@ -2,7 +2,7 @@
 title: "An example with simulated data"
 output: rmarkdown::html_vignette
 vignette: >
-  %\VignetteIndexEntry{An_example_with_simulated_data}
+  # %\VignetteIndexEntry{An example with simulated data}
   %\VignetteEngine{knitr::rmarkdown}
   %\VignetteEncoding{UTF-8}
 ---
@@ -31,7 +31,7 @@ catch_simulated
 
 1. The first step of the `sspm` workflow is to create a `sspm_boundary` from an `sf` object, providing the `boundary` that delineates the boundary regions. The object can then be plotted with `spm_plot` (as can most `sspm` objects). 
 
-```{r boundaries}
+```{r boundaries, fig.width=7, fig.height=5, fig.fullwidth=TRUE, fig.align='center'}
 bounds <- spm_as_boundary(boundaries = sfa_boundaries, 
                           boundary = "sfa")
 
@@ -86,7 +86,7 @@ bounds_delaunay
 
 5. Plotting the object shows the polygons that have been created.
 
-```{r}
+```{r, fig.width=7, fig.height=5, fig.fullwidth=TRUE, fig.align='center'}
 plot(bounds_voronoi)
 ```
 
@@ -121,12 +121,12 @@ biomass_smooth
 
 8. The smoothed results for any smoothed variables (listed in "smoothed vars" above) can be easily plotted:
 
-```{r}
+```{r, fig.width=7, fig.height=5, fig.fullwidth=TRUE, fig.align='center'}
 plot(biomass_smooth, var = "weight_per_km2", log = FALSE, interval = T)
 ```
 You can also make a spatial plot
 
-```{r}
+```{r, fig.width=7, fig.height=5, fig.fullwidth=TRUE, fig.align='center'}
 plot(biomass_smooth, var = "weight_per_km2", use_sf = TRUE)
 ```
 
@@ -211,7 +211,7 @@ sspm_model_fit
 
 16. Plotting the object produces a actual vs predicted plot (with TEST/TRAIN data highlighted.
 
-```{r}
+```{r, fig.width=7, fig.height=5, fig.fullwidth=TRUE, fig.align='center'}
 plot(sspm_model_fit, train_test = TRUE, scales = "free")
 ```
 
@@ -236,17 +236,17 @@ head(biomass_one_step)
 
 18. We can produce an array of plots, as timeseries or as spatial plots
 
-```{r}
+```{r, fig.width=7, fig.height=5, fig.fullwidth=TRUE, fig.align='center'}
 plot(sspm_model_fit, log = T, scales = 'free')
 plot(sspm_model_fit, log = T, use_sf = TRUE)
 ```
 
-```{r}
+```{r, fig.width=7, fig.height=5, fig.fullwidth=TRUE, fig.align='center'}
 plot(sspm_model_fit, biomass = "weight_per_km2_borealis",  scales = "free")
 plot(sspm_model_fit, biomass = "weight_per_km2_borealis", use_sf = TRUE)
 ```
 
-```{r}
+```{r, fig.width=7, fig.height=5, fig.fullwidth=TRUE, fig.align='center'}
 plot(sspm_model_fit, biomass = "weight_per_km2_borealis",
      next_ts = TRUE, aggregate = TRUE, scales = "free", 
      smoothed_biomass = TRUE, interval = T)
