### 2022-12-07 pedersen-fisheries-lab/sspm#125 @VLucet
link to the research doc: https://www.dfo-mpo.gc.ca/csas-sccs/Publications/ResDocs-DocRech/2022/2022_062-eng.html
