### 2022-12-07 #4724 @VLucet
Hi all, my apologies for underestimating the amount of time it is taking me to address all the comments. I wanted to thank again @kellijohnson-NOAA and @quang-huynh  for the detailed and helpful comments (the list of tasks they have given and the linked commits for addressing them are here: https://github.com/pedersen-fisheries-lab/sspm/issues/54 and https://github.com/pedersen-fisheries-lab/sspm/issues/59).

@quang-huynh  I also wanted to let you know that the research document is finally out [here](https://www.dfo-mpo.gc.ca/csas-sccs/Publications/ResDocs-DocRech/2022/2022_062-eng.html), in case there is more you wanted to add to your comment here: https://github.com/pedersen-fisheries-lab/sspm/issues/125)
