### 2022-12-08 ee1d2b3
add RTriangle for cran
