+++ DESCRIPTION
@@ -33,7 +33,8 @@ Imports:
     dplyr,
     purrr,
     stringr,
-    rlang
+    rlang,
+    RTriangle
 Suggests: 
     testthat (>= 3.0.0),
     covr,
