### 2022-12-08 #4724 @quang-huynh
@VLucet Thanks for the link to the research document. The closest alternative to sspm appears to be VAST for spatial surplus production modeling: https://james-thorson-noaa.github.io/docs/tutorials/surplus-production/ The associated paper is https://doi.org/10.1111/1365-2664.12664 which the authors could use for comparison

The major benefit of sspm over VAST appears to be run time, the longer model runtime is VAST is associated with estimating the spatial random field. I've also found a big learning curve is needed with VAST. sspm addresses a much narrower use case which makes it easier to learn for a new user.
