### 2023-01-29 7e0f997
move code for checking if already discretized
