+++ R/spm_discretize.R
@@ -130,6 +130,11 @@ setMethod(f = "spm_discretize",
                     method = "discretization_method"),
           function(boundary_object, method, with, ...) {
 
+            if (checkmate::test_class(boundary_object,
+                                      "sspm_discrete_boundary")) {
+              stop(" Boundary is already discretized")
+            }
+
             if (checkmate::test_class(with, "sspm_dataset")){
               with <- spm_data(with)
             } else if (!is.null(with)) {
@@ -168,16 +173,5 @@ setMethod(f = "spm_discretize",
                   points = discrete[["points"]])
 
             return(new_sspm_discrete_boundary)
-          }
-)
-
-# RE-discretization not allowed for now
-#' @rdname spm_discretize
-#' @export
-setMethod(f = "spm_discretize",
-          signature(boundary_object = "sspm_discrete_boundary"),
-          function(boundary_object, method, with, ...) {
-
-            cli::cli_alert_danger(paste0(" Boundary is already discretized"))
           }
 )
