+++ tests/testthat/test-plot.R
@@ -3,6 +3,4 @@
 test_that("Plotting works as expected", {
   skip("TODO")
 
-  # TODO
-
 })
