+++ NAMESPACE
@@ -1,7 +1,6 @@
 # Generated by roxygen2: do not edit by hand
 
 export("%>%")
-export("dataset<-")
 export("formula_type<-")
 export("formula_vars<-")
 export("is_fitted<-")
@@ -33,7 +32,6 @@ export("spm_time<-")
 export("spm_unique_ID<-")
 export("translated_formula<-")
 export(as_discretization_method)
-export(dataset)
 export(formula_type)
 export(formula_vars)
 export(is_fitted)
@@ -41,7 +39,6 @@ export(is_mapped)
 export(is_split)
 export(method_func)
 export(plot)
-export(plot_trajectory)
 export(predict)
 export(predict_intervals)
 export(raw_formula)
@@ -87,7 +84,6 @@ export(tesselate_voronoi)
 export(translated_formula)
 export(triangulate_delaunay)
 exportMethods("$")
-exportMethods("dataset<-")
 exportMethods("formula_type<-")
 exportMethods("formula_vars<-")
 exportMethods("is_fitted<-")
@@ -119,7 +115,6 @@ exportMethods("spm_time<-")
 exportMethods("spm_unique_ID<-")
 exportMethods("translated_formula<-")
 exportMethods(as_discretization_method)
-exportMethods(dataset)
 exportMethods(formula_type)
 exportMethods(formula_vars)
 exportMethods(is_fitted)
@@ -127,7 +122,6 @@ exportMethods(is_mapped)
 exportMethods(is_split)
 exportMethods(method_func)
 exportMethods(plot)
-exportMethods(plot_trajectory)
 exportMethods(raw_formula)
 exportMethods(smooth_lag)
 exportMethods(smooth_space)
