+++ man/extract-methods.Rd
@@ -6,7 +6,6 @@
 \alias{$,sspm_discrete_boundary-method}
 \alias{$,sspm_dataset-method}
 \alias{$,sspm-method}
-\alias{$,sspm_fit-method}
 \title{Extract methods}
 \usage{
 \S4method{$}{sspm_boundary}(x, name)
@@ -16,8 +15,6 @@
 \S4method{$}{sspm_dataset}(x, name)
 
 \S4method{$}{sspm}(x, name)
-
-\S4method{$}{sspm_fit}(x, name)
 }
 \arguments{
 \item{x}{\strong{[sspm_...]} An object from this package.}
