+++ tests/testthat/test-plot.R
@@ -1,6 +1,0 @@
-# Simple tests for plotting functions
-
-test_that("Plotting works as expected", {
-  skip("TODO")
-
-})
