+++ tests/testthat/test-extracts.R
@@ -7,6 +7,5 @@ test_that("Extractors work as expected", {
   expect_class(biomass_dataset$sfa, "sf")
   expect_class(biomass_dataset_smoothed$sfa, "sf")
   expect_class(sspm_model$sfa, "sf")
-  expect_class(sspm_fit$sfa, "sf")
 
 })
