### 2023-01-29 43718c9
add tests for units closes #57 and closes #30
