+++ man/summary.Rd
@@ -1,0 +1,28 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/summary.R
+\name{summary}
+\alias{summary}
+\alias{summary.sspm}
+\alias{summary,sspm_fit-method}
+\title{Summarises \code{sspm_fit} objects}
+\usage{
+\S4method{summary}{sspm_fit}(object, biomass = NULL)
+}
+\arguments{
+\item{object}{\strong{[sspm_...]} An object from this package.}
+
+\item{biomass}{\strong{[character]} Biomass variable.}
+}
+\value{
+A summary.
+}
+\description{
+Summarises a \code{sspm_fit} object, both in terms of productivity and biomass.
+}
+\examples{
+\dontrun{
+summary(sspm_model_fit)
+summary(sspm_model_fit, biomass = "weight_per_km2_borealis")
+}
+
+}
