### 2023-01-30 #4724 @VLucet
Hi @Bisaloo ! Happy new year as well. @eric-pedersen and I have been discussing how to address the last couple of comment relating to the text (https://github.com/pedersen-fisheries-lab/sspm/issues/125 and https://github.com/pedersen-fisheries-lab/sspm/issues/126). We need a little more time to finalize that. All the other comments and technical issues have been addressed in https://github.com/pedersen-fisheries-lab/sspm/pull/58.
