+++ paper/paper.bib
@@ -1,3 +1,47 @@
+@book{rueGaussianMarkovRandom2005,
+  title = {Gaussian {{Markov}} random fields: theory and applications},
+  author = {Rue, Havard and Held, Leonhard},
+  year = {2005},
+  publisher = {{CRC press}},
+  address = {{New York}}
+}
+
+
+@misc{anderson2022sdmTMB,
+  title = {{{sdmTMB}}: an {{R}} package for fast, flexible, and user-friendly generalized linear mixed effects models with spatial and spatiotemporal random fields},
+  shorttitle = {{{sdmTMB}}},
+  author = {Anderson, Sean C. and Ward, Eric J. and English, Philina A. and Barnett, Lewis A. K.},
+  year = {2022},
+  month = may,
+  primaryclass = {New Results},
+  pages = {2022.03.24.485545},
+  publisher = {{bioRxiv}},
+  doi = {10.1101/2022.03.24.485545},
+  urldate = {2023-05-01},
+  abstract = {Geostatistical data\textemdash spatially referenced observations related to some continuous spatial phenomenon\textemdash{} are ubiquitous in ecology and can reveal ecological processes and inform management decisions. However, appropriate models to analyze these data, such as generalized linear mixed effects models (GLMMs) with Gaussian random fields, are often computationally intensive and challenging to implement, interpret, and evaluate.Here, we introduce the R package sdmTMB, which implements predictive-process SPDE- (stochastic partial differential equation) based spatial and spatiotemporal models. Estimation is conducted via maximum marginal likelihood with Template Model Builder (TMB) but can be extended to penalized likelihood or Bayesian inference. We describe the statistical model, illustrate the package's use through two case studies, and compare the functionality, speed, and interface to related software.We highlight advantages of using sdmTMB for this class of models: (1) sdmTMB provides a flexible interface familiar to users of glm(), lme4, glmmTMB, or mgcv; (2) estimation is often faster than alternatives; (3) sdmTMB provides simple out-of-sample cross validation; (4) non-stationary processes (time-varying and spatially varying coefficients) are easily constructed with a formula interface; and (5) sdmTMB includes features not available as a combination in related packages (e.g., delta/hurdle models, penalized smoothers and break-point effects, anisotropy, abundance index standardization).We hope that sdmTMB's user-friendly interface will open this useful class of models to a wider audience within species distribution modelling and beyond.},
+  archiveprefix = {bioRxiv},
+  chapter = {New Results},
+  copyright = {\textcopyright{} 2022, Posted by Cold Spring Harbor Laboratory. This pre-print is available under a Creative Commons License (Attribution 4.0 International), CC BY 4.0, as described at http://creativecommons.org/licenses/by/4.0/},
+  langid = {english}
+}
+
+
+
+@article{thorson2019Guidance,
+  title = {Guidance for decisions using the {{Vector Autoregressive Spatio-Temporal}} ({{VAST}}) package in stock, ecosystem, habitat and climate assessments},
+  author = {Thorson, James T.},
+  year = {2019},
+  month = feb,
+  journal = {Fisheries Research},
+  volume = {210},
+  pages = {143--161},
+  issn = {0165-7836},
+  doi = {10.1016/j.fishres.2018.10.013},
+  urldate = {2023-05-01},
[…]
