+++ paper/paper.md
@@ -81,17 +81,19 @@ The key workflow steps are:
 ![The sspm workflow.\label{fig:workflow}](../man/figures/flowchart.png){width=90%}
 
 
-# Connections to other spatiotemporal stock assessment approaches
+# Connections to other surplus-production based stock assessment approaches
 
-The general approach used by `sspm` of using a statistical model to estimate spatiotemporally varying population varying biomass indices is closely connected to approaches used by other modern model-based spatial abundance estimation software, such as the `VAST` R package [@thorson2019Guidance] and the `sdmTMB` R package [@anderson2022sdmTMB]. Our method shares the same approach as both `VAST` and `sdmTMB` of using spatially explicit models to estimate local biomass density (\autoref{fig:workflow} steps 1,2, and 4), then aggregating up from those models to predict aggregate stock-level metrics such as total biomass and productivity (\autoref{fig:workflow} steps 8). The multiplicative surplus production model used by `sspm` is also similar to the vector-autoregessive model for biomass changes used by these two packages, as both `VAST` and `sdmTMB` can model local temporal changes as autoregressive processes on the link-scale of a generalized linear model [@thorson2019Guidance]. 
+The `sspm` package uses a model-based, random-effects based approach to estimate the effects of ecosystem drivers on surplus production across space and time. Our approach is conceptually related to the stochastic stock assessment approaches used by the R packages `spict` [@pedersen_stochastic_2017] and `jabba` [@winker_jabba_2018] R packages for surplus production modelling, in that we assume that biomass dynamics can be modelled as effectively a logistic growth model with both process and measurement error. While `sspm` does not currently have the capacity to model biomass dynamics as a continuous-time process, as with `spict`, or incorporate prior parameter information on catchability or biomass dynamics as in `jabba`, `sspm` can model spatially and temporally varying productivity, which is currently not possible in these models.
 
-One major difference between the `sspm` package and other model-based spatiotemporal modelling packages is its special-purpose nature. The `spm_smooth` function uses a computationally simpler (although somewhat less flexible) Conditional Autoregressive (CAR) model [@rueGaussianMarkovRandom2005] for modelling spatial variation in covariates and biomass, as compared to the more complex spatial random effects possible with `VAST` and `sdmTMB`; this has the advantage of computational speed and less user knowledge of how to set up complex spatial grids, although it is less flexible. This means that `sspm` should be easier to adapt to novel fisheries than more complex packages that require more user modelling knowledge. 
+The `sspm` package can be viewed as a spatiotemporal Model of Intermediate Complexity  [a 'MICE-in-space' model; @thorson_spatio-temporal_2019] that can incorporate effects of other species and ecosystem drivers as well as changes in fishing pressure on stock status. Our approach is closely connected to approaches used by other modern model-based spatial abundance estimation software, such as the `VAST` R package [@thorson_spatio-temporal_2019] and the `sdmTMB` R package [@anderson2022sdmTMB]. Our method shares the same approach as both `VAST` and `sdmTMB` of using spatially explicit models to estimate local biomass density (\autoref{fig:workflow} steps 1,2, and 4), then aggregating up from those models to predict aggregate stock-level metrics such as total biomass and productivity (\autoref{fig:workflow} steps 8). The multiplicative surplus production model used by `sspm` is also conceptually similar to the vector-autoregessive model for biomass changes used by these two packages, as both `VAST` and `sdmTMB` can model local temporal changes as autoregressive processes on the link-scale of a generalized linear model. The `sspm` package cannot, however, model the dynamics of multiple species simultaneously; multi-species modelling would require generating a separate surplus production model (\autoref{fig:workflow} steps 5 and 6) for each species of interest.
 
[…]
