+++ paper/paper.bib
@@ -1,3 +1,22 @@
+@article{thorson_spatio-temporal_2019,
+	title = {Spatio-temporal models of intermediate complexity for ecosystem assessments: {A} new tool for spatial fisheries management},
+	volume = {20},
+	copyright = {© 2019 John Wiley \& Sons Ltd},
+	issn = {1467-2979},
+	shorttitle = {Spatio-temporal models of intermediate complexity for ecosystem assessments},
+	url = {https://onlinelibrary.wiley.com/doi/abs/10.1111/faf.12398},
+	doi = {https://doi.org/10.1111/faf.12398},
+	language = {en},
+	number = {6},
+	urldate = {2021-03-30},
+	journal = {Fish and Fisheries},
+	author = {Thorson, James T. and Adams, Grant and Holsman, Kirstin},
+	year = {2019},
+	note = {\_eprint: https://onlinelibrary.wiley.com/doi/pdf/10.1111/faf.12398},
+	pages = {1083--1099},
+}
+
+
 @book{rueGaussianMarkovRandom2005,
   title = {Gaussian {{Markov}} random fields: theory and applications},
   author = {Rue, Havard and Held, Leonhard},
@@ -298,10 +317,11 @@
 @techreport{pedersenNewSpatialEcosystembased2021,
 	type = {Can. {Sci}. {Advis}. {Sec}. {Res}. {Doc}.},
 	title = {A new spatial ecosystem-based surplus production model for {SFA} 4-6 {Northern} {Shrimp}},
-	number = {2022/nn},
+	number = {2022/062},
 	author = {Pedersen, Eric J. and Skanes, Katherine and Le Corre, Nicolas and Koen-Alonso, Mariano and Baker, Krista},
 	year = {2022},
-	pages = {in press},
+	pages = {1-64},
+	url = {https://publications.gc.ca/collections/collection_2022/mpo-dfo/fs70-4/Fs70-4-2022-012-eng.pdf},
 }
 
 
