+++ paper/paper.bib
@@ -1,3 +1,36 @@
+
+@article{pedersen_stochastic_2017,
+	title = {A stochastic surplus production model in continuous time},
+	volume = {18},
+	issn = {1467-2979},
+	url = {https://onlinelibrary.wiley.com/doi/abs/10.1111/faf.12174},
+	doi = {10.1111/faf.12174},
+	language = {en},
+	number = {2},
+	urldate = {2023-05-02},
+	journal = {Fish and Fisheries},
+	author = {Pedersen, Martin W and Berg, Casper W},
+	year = {2017},
+	pages = {226--243},
+}
+
+@article{winker_jabba_2018,
+	title = {{JABBA}: {Just} {Another} {Bayesian} {Biomass} {Assessment}},
+	volume = {204},
+	issn = {0165-7836},
+	shorttitle = {{JABBA}},
+	url = {https://www.sciencedirect.com/science/article/pii/S0165783618300845},
+	doi = {10.1016/j.fishres.2018.03.010},
+	language = {en},
+	urldate = {2023-05-02},
+	journal = {Fisheries Research},
+	author = {Winker, Henning and Carvalho, Felipe and Kapur, Maia},
+	month = aug,
+	year = {2018},
+	pages = {275--288},
+}
+
+
 @article{thorson_spatio-temporal_2019,
 	title = {Spatio-temporal models of intermediate complexity for ecosystem assessments: {A} new tool for spatial fisheries management},
 	volume = {20},
