+++ paper/paper.bib
@@ -125,7 +125,8 @@
 	editor = {Maravelias, Christos},
 	month = apr,
 	year = {2019},
-	pages = {fsz048},
+	pages = {1305--1315},
+	volume = {76},
 	file = {Karp et al. - 2019 - Accounting for shifting distributions and changing.pdf:/home/vlucet/Zotero/storage/JP6BDALA/Karp et al. - 2019 - Accounting for shifting distributions and changing.pdf:application/pdf},
 }
 
@@ -242,10 +243,12 @@
 
 @article{prager_suite_1994,
 	title = {A suite of extensions to a nonequilibrium surplus-production},
+	journal = {Fishery Bulletin - National Oceanic and Atmospheric Administration},
 	language = {en},
 	author = {Prager, Michael H},
 	year = {1994},
-	pages = {17},
+	volume = {92},
+	pages = {374--389},
 	file = {Prager - A suite of extensions to a nonequilibrium surplus-.pdf:/home/vlucet/Zotero/storage/4HHXRZ4Y/Prager - A suite of extensions to a nonequilibrium surplus-.pdf:application/pdf},
 }
 
@@ -297,7 +300,6 @@
 	file = {Snapshot:/home/vlucet/Zotero/storage/L24AL4DI/a-primer-of-ecology-9780878933181.html:text/html},
 }
 
-
 @article{thorson_importance_2015,
 	title = {The importance of spatial models for estimating the strength of density dependence},
 	volume = {96},
@@ -312,7 +314,6 @@
 	year = {2015},
 	pages = {1202--1212},
 }
-
 
 @book{walters_fisheries_2004,
 	address = {Princeton, N.J., U.S.A.},
@@ -340,7 +341,7 @@
 @techreport{dfo_assessment_2019,
 	type = {techreport},
 	institution = {Canadian Science Advisory Secretariat (CSAS)},
-	title = {An assessment of {Northern} {Shrimp} (\textit{{Pandalus} borealis}) in {Shrimp} {Fishing} {Areas} 4–6 and of {Striped} {Shrimp} (\textit{{Pandalus} montagui}) in shrimp fishing area 4 in 2018},
+	title = {An Assessment of {Northern} {Shrimp} (Pandalus borealis) in {Shrimp} {Fishing} {Areas} 4–6 and of {Striped} {Shrimp} (Pandalus montagui) in Shrimp Fishing Area 4 in 2018},
 	number = {2019/027},
 	author = {{DFO}},
 	year = {2019},
@@ -348,16 +349,12 @@
 }
 
 @techreport{pedersenNewSpatialEcosystembased2021,
-	type = {Can. {Sci}. {Advis}. {Sec}. {Res}. {Doc}.},
-	title = {A new spatial ecosystem-based surplus production model for {SFA} 4-6 {Northern} {Shrimp}},
-	number = {2022/062},
+	type = {Canadian Science Advisory Secretariat (CSAS) Research Document},
+	title = {A New Spatial Ecosystem-Based Surplus Production Model for {SFA} 4-6 {Northern} {Shrimp}},
 	author = {Pedersen, Eric J. and Skanes, Katherine and Le Corre, Nicolas and Koen-Alonso, Mariano and Baker, Krista},
 	year = {2022},
-	pages = {1-64},
-	url = {https://publications.gc.ca/collections/collection_2022/mpo-dfo/fs70-4/Fs70-4-2022-012-eng.pdf},
+	url = {https://www.dfo-mpo.gc.ca/csas-sccs/Publications/ResDocs-DocRech/2022/2022_062-eng.html}
 }
-
-
 
 @article{plaganyi2014Multispecies,
 	title = {Multispecies fisheries management and conservation: tactical applications using models of intermediate complexity},
