### 2023-05-16 #4724 @quang-huynh
The revision is quite thorough in the comparison of methods and software packages. It was helpful for me and the text is appreciated. The paper looks good to me, pending a DOI fix in Line 223.

The figures in the vignette have also been revised and looked good to me.

The Github repository just needs a section on: clear guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support
