### 2023-05-16 #4724 @quang-huynh
One more thing: the package can be used for more than just northern shrimp fisheries. Do the authors want to update the title to broaden the application? For example: "The sspm R package: spatial surplus production models for fisheries management" or "The sspm R package for spatially-explicit surplus production population models"
