### 2023-05-25 #4724 @Bisaloo
Thanks for the changes @VLucet!

@quang-huynh @kellijohnson-NOAA, could you review your respective checklists ([1](https://github.com/openjournals/joss-reviews/issues/4724#issuecomment-1238431652), [2](https://github.com/kellijohnson-NOAA)) and verify if the unchecked items have been addressed by the latest changes please? Thanks a lot for your patience and your work on this :pray:
