### 2023-05-25 pedersen-fisheries-lab/sspm#131 @kellijohnson-NOAA
Change "Productivity" to "Production"
"Surplus production models" is the typical phrase used in fisheries science but in this paragraph you have "Surplus Productivity Models". I recommend changing all instances throughout the package and the manuscript to my suggestion.
https://github.com/pedersen-fisheries-lab/sspm/blob/faff4a5ac3cb9f4eb2a26b2666e4b89a06679f27/paper/paper.md?plain=1#L31

This issue is part of a review for JOSS, see https://github.com/openjournals/joss-reviews/issues/4724.
