### 2023-05-25 pedersen-fisheries-lab/sspm#132 @kellijohnson-NOAA
"fisheries stock models" is not a common term
I suggest changing "fisheries stock models" to "fisheries assessment models"
https://github.com/pedersen-fisheries-lab/sspm/blob/faff4a5ac3cb9f4eb2a26b2666e4b89a06679f27/paper/paper.md?plain=1#L37

This issue is part of a review for JOSS, see https://github.com/openjournals/joss-reviews/issues/4724.
