### 2023-05-25 pedersen-fisheries-lab/sspm#133 @kellijohnson-NOAA
Change "global warming" to "climate change"
Global warming suggests that all locations will increase in temperature where climate change acknowledges that some places will experience situations where the temperature decreases or that changes related to climate but not specific to a temperature are also slated to change.

https://github.com/pedersen-fisheries-lab/sspm/blob/faff4a5ac3cb9f4eb2a26b2666e4b89a06679f27/paper/paper.md?plain=1#L45

This issue is part of a review for JOSS, see https://github.com/openjournals/joss-reviews/issues/4724.
