### 2023-05-25 pedersen-fisheries-lab/sspm#134 @kellijohnson-NOAA
Confusing sentence
- [x] You do not need to redefine SSPM because you already defined the acronym on the previous line, just use the acronym here.
- [x] "The basic model it implements was first" is confusing because upon first read I was not sure what "it" was referring to.

https://github.com/pedersen-fisheries-lab/sspm/blob/faff4a5ac3cb9f4eb2a26b2666e4b89a06679f27/paper/paper.md?plain=1#L55

This issue is part of a review for JOSS, see https://github.com/openjournals/joss-reviews/issues/4724.
