### 2023-05-25 pedersen-fisheries-lab/sspm#137 @kellijohnson-NOAA
Is step 7 "fit()" or "spm()" b/c the diagram says the former
but here
https://github.com/pedersen-fisheries-lab/sspm/blob/faff4a5ac3cb9f4eb2a26b2666e4b89a06679f27/paper/paper.md?plain=1#L78
you say the latter and both functions are available in the package.

This issue is part of a review for JOSS, see https://github.com/openjournals/joss-reviews/issues/4724.
