### 2023-05-25 pedersen-fisheries-lab/sspm#140 @kellijohnson-NOAA
Grammar edits to line 88
- [x] Change "by the R packages `spict` [@pedersen_stochastic_2017] and `jabba` [@winker_jabba_2018] R packages for surplus production modelling," to "by `spict` [@pedersen_stochastic_2017] and `jabba` [@winker_jabba_2018], R packages for surplus production modelling,".
- [x] Place a comma before "as in `jabba`".

https://github.com/pedersen-fisheries-lab/sspm/blob/faff4a5ac3cb9f4eb2a26b2666e4b89a06679f27/paper/paper.md?plain=1#L88

This issue is part of a review for JOSS, see https://github.com/openjournals/joss-reviews/issues/4724.
