### 2023-05-25 pedersen-fisheries-lab/sspm#140 @VLucet
done
