### 2023-05-25 pedersen-fisheries-lab/sspm#142 @kellijohnson-NOAA
Crowder et al. is not a book
I am not sure why you embrace every word in the title of Crowder et al. because it is not a book. Shouldn't you just let the csl file deal with how it wants the title to be capitalized rather than forcing each word to be upper cased?

https://github.com/pedersen-fisheries-lab/sspm/blob/72993d4bc6319e198335624d2057ad966d70e1c9/paper/paper.bib#L208

This issue is part of a review for JOSS, see https://github.com/openjournals/joss-reviews/issues/4724.
