### 2023-05-25 pedersen-fisheries-lab/sspm#142 @VLucet
This is not voluntary. Zotero extracts the bib info from he Annual reviews website and exports it as such, because paper titles tend to be capitalized. I can change it.
