### 2023-05-25 pedersen-fisheries-lab/sspm#142 @kellijohnson-NOAA
Ahh thanks for the information regarding your process ... it is often the case that publishers use incorrect LaTeX or formatting rules in their citation information. Another typo I found was with prager_suite_1994 where the journal should just be "Fishery Bulletin" with the NOAA stuff. I actually work for NOAA, which is why I noticed it.
