### 2023-05-25 pedersen-fisheries-lab/sspm#142 @VLucet
For sure, ill fix that too. Same thing there, Zotero generated the name.
