+++ paper/paper.bib
@@ -243,7 +243,7 @@
 
 @article{prager_suite_1994,
 	title = {A suite of extensions to a nonequilibrium surplus-production},
-	journal = {Fishery Bulletin - National Oceanic and Atmospheric Administration},
+	journal = {Fishery Bulletin},
 	language = {en},
 	author = {Prager, Michael H},
 	year = {1994},
