### 2023-05-31 #4724 @kellijohnson-NOAA
Sorry @VLucet that you had to ask. I have just one remaining check-list item that I have not decided if it is complete or not. You have many tests using testthat, which I think are great. But, I am wondering if you have a test that ensures the statistical properties of the answer. For example, if you were running a linear model you could test the parameter estimates against the output from `lm()` to ensure that the statistical properties of the model are implemented correctly. I am uncertain if you are doing that here or not and even how you would go about doing it for a complicated model. Nevertheless, I think at a bare minimum you could test a super simple example without spatial structure to ensure that the underlying code is working as planned. I have yet to be able to find anything like this in {sspm}. If the tests already exist, can you please point me to them? Thanks 🙏.
