### 2023-05-31 #4724 @VLucet
Hi @kellijohnson-NOAA. The closest [test we have](https://github.com/pedersen-fisheries-lab/sspm/blob/main/tests/testthat/test-fit.R) is one that ensures the fitting of smooths using our fake dataset consistently outputs the "expected intercept", to ensure that if there was a change in how mgcv fits things we would notice the departure from that value. But it's just an "anchor point" for the reproducibility of results.

I understand your question as a general question about correctness of the statistical answer. I would find it difficult to test that programmatically for such a model. The model is using multiple building blocks from mgcv: the approach I tried to take for testing is to make sure those individual building blocks (for example, the penalty matrices) are accurately generated and put together. But ultimately, I am relying on mgcv's extensive testing (and our general understanding of how those building blocks should fit together, ie, the statistical approach we defend in the paper) to trust that those building blocks will indeed work well together and deliver the promised approach. The objects the package creates, are, ultimately, mgcv objects which can be inspected and checked (formulas, penalty matrices, etc..). Given that this package mainly wraps mgcv, I would say the answer to your questions is the ultimately the same one as "how to we make sure a GAM is statiscally correct?", which is out of scope for this package I would think.

I'll call on @eric-pedersen, to add to this, and maybe correct/clarify.
