### 2023-05-31 #4724 @eric-pedersen
Hi @kellijohnson-NOAA. Thanks for the feedback. This is definitely an important consideration to think about.

As @VLucet mentioned, validating "statistical properties" is somewhat difficult given the breadth of possible models that can be fit in this framework, and the complexities of `mgcv`.

I could see adding a couple tests here to ensure that `spm_smooth` and `spm` produce the same estimated GAM coefficients and penalties as using the equivalent direct calls to `gam` with the same data; that way, we can show that the constructor functions (`smooth_space`, `smooth_time`, etc.) are producing the correct formula, and it will double as additional documentation: showing what GAM models `spm_smooth` and `spm` are estimating.

Would that work as a solution for this?
