### 2023-06-08 #4724 @Bisaloo
Yes, it looks good :+1:! Could you please:

- add the authors' ORCIDs on zenodo
- fix the license on zenodo to show MIT rather than "Other (open)"

And please let me know if you've checked the other box in the checklist.
