### 2023-06-19 #4724 @VLucet
@Bisaloo I think what I am confused about is that @quang-huynh has been made author on Zenodo (as it is to be expected as they contributed to the software via PR). Does that mean the authors on the paper (not on the software package) have to be updated to reflect that, or is it okay to have separate author lists? I am not opposed to adding a co-author to the paper (@eric-pedersen can chime in if otherwise), I am just asking what is the usual etiquette in this situation.
