### 2023-06-19 #4724 @Bisaloo
It is possible to edit the metadata on zenodo, which is what we usually do to reflect the authors on the paper. If you are not able to do it, it's not a big deal, we can wave this requirement, which is why I didn't mention it in my last message.
