### 2023-06-20 #4724 @VLucet
I see the error, the ref `thorson_importance_2015` is defined twice. I'll fix it now.
