### 2023-06-20 d50bd28
delete cran sub file, fix duplicate rep in paper
