+++ paper/paper.bib
@@ -300,21 +300,6 @@
 	file = {Snapshot:/home/vlucet/Zotero/storage/L24AL4DI/a-primer-of-ecology-9780878933181.html:text/html},
 }
 
-@article{thorson_importance_2015,
-	title = {The importance of spatial models for estimating the strength of density dependence},
-	volume = {96},
-	issn = {1939-9170},
-	url = {https://onlinelibrary.wiley.com/doi/abs/10.1890/14-0739.1},
-	doi = {10.1890/14-0739.1},
-	language = {en},
-	number = {5},
-	urldate = {2021-12-01},
-	journal = {Ecology},
-	author = {Thorson, James T. and Skaug, Hans J. and Kristensen, Kasper and Shelton, Andrew O. and Ward, Eric J. and Harms, John H. and Benante, James A.},
-	year = {2015},
-	pages = {1202--1212},
-}
-
 @book{walters_fisheries_2004,
 	address = {Princeton, N.J., U.S.A.},
 	title = {Fisheries ecology and management},
