+++ CITATION.cff
@@ -1,15 +1,34 @@
-cff-version: 1.1.0
-message: "If you use this software, please cite it as below."
+cff-version: "1.2.0"
 authors:
+- family-names: Lucet
+  given-names: Valentin
+  orcid: "https://orcid.org/0000-0003-0268-818X"
+- family-names: Pedersen
+  given-names: Eric J.
+  orcid: "https://orcid.org/0000-0003-1016-540X"
+doi: 10.5281/zenodo.8015102
+message: If you use this software, please cite our article in the
+  Journal of Open Source Software.
+preferred-citation:
+  authors:
   - family-names: Lucet
     given-names: Valentin
-    orcid: https://orcid.org/0000-0003-0268-818X
+    orcid: "https://orcid.org/0000-0003-0268-818X"
   - family-names: Pedersen
-    given-names: Eric
-    orcid: https://orcid.org/0000-0003-1016-540X
-  - family-names: Huynh
-    given-names: Quang
-    orcid: https://orcid.org/0000-0001-7835-4376
-title: pedersen-fisheries-lab/sspm: sspm v1.0.0 Zenodo + JOSS release
-version: v1.0.0
-date-released: 2023-06-07
+    given-names: Eric J.
+    orcid: "https://orcid.org/0000-0003-1016-540X"
+  date-published: 2023-06-22
+  doi: 10.21105/joss.04724
+  issn: 2475-9066
+  issue: 86
+  journal: Journal of Open Source Software
+  publisher:
+    name: Open Journals
+  start: 4724
+  title: The sspm R package for spatially-explicit surplus production
+    population models
+  type: article
+  url: "https://joss.theoj.org/papers/10.21105/joss.04724"
+  volume: 8
+title: The sspm R package for spatially-explicit surplus production
+  population models
